+++ lib/lazy_high_charts.rb
@@ -1,8 +1,7 @@
-require 'lazy_high_charts/high_chart'
-require 'lazy_high_charts/layout_helper.rb'
-
+require File.join(File.dirname(__FILE__), *%w[lazy_high_charts high_chart])
+require File.join(File.dirname(__FILE__), *%w[lazy_high_charts layout_helper]) 
 require File.join(File.dirname(__FILE__), *%w[lazy_high_charts railtie]) if defined?(::Rails::Railtie)
 
 module LazyHighCharts
-  # Your code goes here...
+
 end